# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/ticket": {
      "post": {
        "operationId": "TicketController_create",
        "summary": "ADMIN: Create a ticket",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTicketDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Create a ticket operation completed successfully."
          },
          "201": {
            "description": "The ticket has been successfully created."
          }
        },
        "tags": [
          "ticket"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "get": {
        "operationId": "TicketController_findAll",
        "summary": "ADMIN: Get all tickets",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Get all tickets operation completed successfully."
          }
        },
        "tags": [
          "ticket"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/ticket/search": {
      "get": {
        "operationId": "TicketController_search",
        "summary": "PUBLIC: Search tickets",
        "parameters": [
          {
            "name": "ticket_type",
            "required": true,
            "in": "query",
            "description": "Type of the ticket",
            "schema": {
              "example": "ROUND_TRIP",
              "enum": [
                "ONE_WAY",
                "ROUND_TRIP"
              ],
              "type": "string"
            }
          },
          {
            "name": "departure_airport_code",
            "required": true,
            "in": "query",
            "description": "Code of the departure airport",
            "schema": {
              "example": "JFK",
              "type": "string"
            }
          },
          {
            "name": "arrival_airport_code",
            "required": true,
            "in": "query",
            "description": "Code of the arrival airport",
            "schema": {
              "example": "LAX",
              "type": "string"
            }
          },
          {
            "name": "outbound_day",
            "required": true,
            "in": "query",
            "description": "Departure date of the flight (YYYY-MM-DD)",
            "schema": {
              "example": "2023-12-01",
              "type": "string"
            }
          },
          {
            "name": "return_day",
            "required": false,
            "in": "query",
            "description": "Return date of the flight (YYYY-MM-DD)",
            "schema": {
              "example": "2023-12-03",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The tickets have been successfully retrieved."
          }
        },
        "tags": [
          "ticket"
        ]
      }
    },
    "/ticket/search-by-outbound-time": {
      "get": {
        "operationId": "TicketController_searchByOutboundTime",
        "summary": "PUBLIC: Search tickets by outbound flight time",
        "parameters": [
          {
            "name": "date",
            "required": true,
            "in": "query",
            "schema": {
              "example": "2023-12-25",
              "type": "string"
            }
          },
          {
            "name": "before",
            "required": true,
            "in": "query",
            "schema": {
              "example": true,
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Find tickets where outbound flight departs before or after a specified date"
          }
        },
        "tags": [
          "ticket"
        ]
      }
    },
    "/ticket/user/{userId}": {
      "get": {
        "operationId": "TicketController_findAllByUserId",
        "summary": "Get all tickets belonging to a user. The user can only access their own tickets",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get all tickets belonging to a user. The user can only access their own tickets operation completed successfully."
          }
        },
        "tags": [
          "ticket"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/ticket/{id}": {
      "get": {
        "operationId": "TicketController_findOne",
        "summary": "ADMIN: Find a ticket by id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Find a ticket by id operation completed successfully."
          }
        },
        "tags": [
          "ticket"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "patch": {
        "operationId": "TicketController_update",
        "summary": "ADMIN: Update a ticket",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTicketDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Update a ticket operation completed successfully."
          }
        },
        "tags": [
          "ticket"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "delete": {
        "operationId": "TicketController_remove",
        "summary": "ADMIN: Delete a ticket",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Delete a ticket operation completed successfully."
          }
        },
        "tags": [
          "ticket"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/ticket/cancel/{id}": {
      "patch": {
        "operationId": "TicketController_cancel",
        "summary": "Cancel a ticket",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cancel a ticket operation completed successfully."
          }
        },
        "tags": [
          "ticket"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/ticket/book/{id}": {
      "patch": {
        "operationId": "TicketController_bookTicket",
        "summary": "Book a ticket",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookTicketDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Book a ticket operation completed successfully."
          },
          "400": {
            "description": "Invalid booking request or ticket already booked."
          }
        },
        "tags": [
          "ticket"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "QAirline",
    "description": "The QAirline API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreateTicketDto": {
        "type": "object",
        "properties": {
          "outbound_flight_id": {
            "type": "number",
            "description": "Outbound flight ID",
            "example": 1
          },
          "return_flight_id": {
            "type": "number",
            "description": "Return flight ID",
            "example": 1
          },
          "user_id": {
            "type": "number",
            "description": "User ID",
            "example": 1
          },
          "booking_date": {
            "format": "date-time",
            "type": "string",
            "description": "Booking date",
            "example": "2021-12-31T23:59:59Z"
          },
          "ticket_type": {
            "type": "string",
            "enum": [
              "ONE_WAY",
              "ROUND_TRIP"
            ],
            "description": "Ticket type",
            "example": "ONE_WAY"
          },
          "booking_class": {
            "type": "string",
            "enum": [
              "ECONOMY",
              "BUSINESS",
              "FIRST_CLASS"
            ],
            "description": "Booking class",
            "example": "ECONOMY"
          },
          "booking_seat_code": {
            "type": "string",
            "description": "Booking seat code",
            "example": "A1"
          },
          "description": {
            "type": "string",
            "description": "Description",
            "example": "This is a ticket"
          },
          "total_passengers": {
            "type": "number",
            "description": "Total passengers",
            "example": 1
          },
          "base_price": {
            "type": "string",
            "description": "Base price",
            "example": "100"
          },
          "booking_status": {
            "type": "string",
            "enum": [
              "PENDING",
              "CONFIRMED",
              "CANCELLED"
            ],
            "description": "Booking status",
            "example": "CONFIRMED"
          }
        },
        "required": [
          "outbound_flight_id",
          "return_flight_id",
          "user_id",
          "booking_date",
          "ticket_type",
          "booking_class",
          "booking_seat_code",
          "description",
          "total_passengers",
          "base_price",
          "booking_status"
        ]
      },
      "BookTicketDto": {
        "type": "object",
        "properties": {
          "booking_seat_code": {
            "type": "string",
            "description": "Seat code for the booking",
            "example": "A12-B15"
          },
          "total_passengers": {
            "type": "number",
            "description": "Number of passengers",
            "example": 2
          }
        },
        "required": [
          "booking_seat_code",
          "total_passengers"
        ]
      },
      "UpdateTicketDto": {
        "type": "object",
        "properties": {
          "outbound_flight_id": {
            "type": "number",
            "description": "Outbound flight ID",
            "example": 1
          },
          "return_flight_id": {
            "type": "number",
            "description": "Return flight ID",
            "example": 1
          },
          "user_id": {
            "type": "number",
            "description": "User ID",
            "example": 1
          },
          "booking_date": {
            "format": "date-time",
            "type": "string",
            "description": "Booking date",
            "example": "2021-12-31T23:59:59Z"
          },
          "ticket_type": {
            "type": "string",
            "enum": [
              "ONE_WAY",
              "ROUND_TRIP"
            ],
            "description": "Ticket type",
            "example": "ONE_WAY"
          },
          "booking_class": {
            "type": "string",
            "enum": [
              "ECONOMY",
              "BUSINESS",
              "FIRST_CLASS"
            ],
            "description": "Booking class",
            "example": "ECONOMY"
          },
          "booking_seat_code": {
            "type": "string",
            "description": "Booking seat code",
            "example": "A1"
          },
          "description": {
            "type": "string",
            "description": "Description",
            "example": "This is a ticket"
          },
          "total_passengers": {
            "type": "number",
            "description": "Total passengers",
            "example": 1
          },
          "base_price": {
            "type": "string",
            "description": "Base price",
            "example": "100"
          },
          "booking_status": {
            "type": "string",
            "enum": [
              "PENDING",
              "CONFIRMED",
              "CANCELLED"
            ],
            "description": "Booking status",
            "example": "CONFIRMED"
          }
        },
        "required": [
          "outbound_flight_id",
          "return_flight_id",
          "user_id",
          "booking_date",
          "ticket_type",
          "booking_class",
          "booking_seat_code",
          "description",
          "total_passengers",
          "base_price",
          "booking_status"
        ]
      },
...
    }
  }
}